Content was misplaced in example folder. Relocated to lsls folder. Program is functioning after test runs.
