+++ lsls/lsls.c
@@ -6,13 +6,32 @@
  */
 int main(int argc, char **argv)
 {
-  // Parse command line
+    int i;
+    struct dirent *pDirent;
+    DIR *pDir;
 
-  // Open directory
+    // printf("There are %d command line argument(s):\n", argc);
 
-  // Repeatly read and print entries
+    // for (i = 0; i < argc; i++) {
+    //     printf("   %s\n", argv[i]);
+    // }
 
-  // Close directory
+    if (argc < 2) {
+        printf("Usage: ./lsls <directory>\n");
+    return 1;
+    }
+
+    pDir = opendir(argv[1]);
+    if (pDir == NULL){
+        printf("Cannot open directory %s\n", argv[1]);
+        return 0;
+    }
+    while ((pDirent = readdir(pDir)) != NULL){
+        printf("%s\n", pDirent -> d_name);
+    }
+    closedir(pDir);
+    
+    return 0;
 
   return 0;
 }
